Move Gluon DSA tests to AMD package
